add auth to neccessary routes

+++ routes/api.php
@@ -5,19 +5,21 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\WebsiteController;
 use Illuminate\Support\Facades\Route;
 
+//->middleware('role:admin')
+
 Route::post("/login", [AuthController::class, 'login']);
-Route::middleware('auth:api')->get("/user", [UserController::class, 'user']);
+Route::get("user", [UserController::class, 'user'])->middleware(['auth:api']);
 
 Route::controller(WebsiteController::class)
     ->prefix("websites")->group(function () {
         Route::get("/", "index");
-        Route::post("/", "store");
-        Route::delete("/{id}", "delete");
-        Route::put("/{id}", "update");
+        Route::post("/", "store")->middleware(['auth:api', 'role:admin']);
+        Route::delete("/{id}", "delete")->middleware(['auth:api', 'role:admin']);
+        Route::put("/{id}", "update")->middleware(['auth:api', 'role:admin']);
     });
 Route::controller(UserController::class)
-    ->prefix("users")->group(function () {
-        Route::get("/admins", "index");
+    ->prefix("users")->middleware(['auth:api', 'role:admin'])->group(function () {
+        Route::get("/admins", "index"); //getAdmins
         Route::get("/users", "getUsers");
         Route::post("/{id}", "store");
         Route::delete("/{id}", "delete");
